pyproject.toml
[project]
name = "bookeditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.118.0",
    "google-api-python-client>=2.169.0",
    "gspread>=6.2.0",
    "langchain>=0.3.25",
    "langchain-google-vertexai>=2.0.21",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.1",
    "oauth2client>=4.1.3",
    "pandas>=2.2.3",
    "pip>=25.1.1",
    "python-dotenv>=1.1.0",
]
